backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
greenlet>=3.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
pydantic[email]>=2.5.3
pydantic-settings>=2.1.0
stripe>=7.12.0
aiosmtplib>=3.0.1
python-dotenv>=1.0.0
httpx>=0.26.0
jinja2>=3.1.3
email-validator>=2.1.0.post1
